// src/app/api/swagger/models/api-order-dto.ts
/* tslint:disable */
/* eslint-disable */
import { ApiAdditionalRequirementDto } from '../models/api-additional-requirement-dto';
import { ApiCustomerContactInfoAltDto } from '../models/api-customer-contact-info';
import { ApiDeliveryMethodAltEnum } from '../models/api-delivery-method';
import { ApiFederalLawLimitAltEnum } from '../models/api-federal-law-limit';
import { ApiOrderDetailDto } from '../models/api-order-detail-dto';
import { ApiOrderDocumentDto } from '../models/api-order-document-dto';
import { ApiOrderStateEnum } from '../models/api-order-state-enum';
import { ApiTradeBusinessFlowEnum } from '../models/api-trade-business-flow-enum';
import { ApiTradeExtCustomerInfoExcelDto } from '../models/api-trade-ext-customer-info-excel-dto';
import { ApiZmoFinanceSourceEnum } from '../models/api-zmo-finance-source-enum';
export interface ApiOrderDto {
  AdditionalRequirements?: Array<ApiAdditionalRequirementDto> | null;
  BusinessFlowType?: ApiTradeBusinessFlowEnum;
  Comment?: string | null;
  ContactInfo?: ApiCustomerContactInfoAltDto | null;
  CreatorUserId?: number | null;
  CurrencyCode?: string | null;
  CustomerId?: string | null;
  DealConclusionTermInfo?: string | null;
  DeliveryMethod?: ApiDeliveryMethodAltEnum;
  DeliveryPlace?: string | null;
  DeliveryTerms?: string | null;
  Description?: string | null;
  ExternalSystemId?: number | null;
  FederalLawLimit?: ApiFederalLawLimitAltEnum | null;
  Id?: number;
  InitialPrice?: number | null;
  IsAutoCompleteDeal?: boolean;
  IsBasedOnPublicOffer?: boolean | null;
  IsCartOrder?: boolean;
  IsContractInElectronicForm?: boolean | null;
  IsCustomPrice?: boolean;
  IsFederalLawsEnable?: boolean;
  IsParentContractNumberNotSpecified?: boolean | null;
  IsParticipantCanCreateDeal?: boolean;
  IsSanctionedPurchase?: boolean | null;
  IsUnitBidding?: boolean;
  KeepGuaranteeForAllParticipants?: boolean;
  KeepGuaranteeForParticipantsCount?: number | null;
  LastModificationDate?: string | null;
  MaxDealPrice?: number | null;
  OrderDetails?: Array<ApiOrderDetailDto> | null;
  OrderDocuments?: Array<ApiOrderDocumentDto> | null;
  OrderName?: string | null;
  OrderState?: ApiOrderStateEnum;
  ParentContractNumber?: string | null;
  ParentNotTookPlaceTradeId?: number | null;
  ParticipantComment?: string | null;
  ParticipantId?: string | null;
  PlanedDealSignDate?: string | null;
  PossibleSendByTransportCompany?: boolean;
  RegionCode?: string | null;
  TradeExtCustomerInfo?: ApiTradeExtCustomerInfoExcelDto | null;
  TradeIdentificationCode?: string | null;
  ZmoFinanceSource?: ApiZmoFinanceSourceEnum | null;
  ZmoFzType?: number | null;
}